Add begin to Prewitt Filter
Add Prewitt Filter, but does not work. It is problem with setting
correct values in matrix.
It worth to think about delete Prewitt filter, because it is very
similar like Sobel filter. maybe better add Scharr filter, which it is
implemented in opencv.
Add change to Sobel filter, now filter works on grayscale image.